+++ services/ordering/src/main/java/com/eshop/ordering/api/application/commands/SetAwaitingValidationOrderStatusCommand.java
@@ -2,5 +2,5 @@ package com.eshop.ordering.api.application.commands;
 
 import an.awesome.pipelinr.Command;
 
-public record SetAwaitingValidationOrderStatusCommand(Integer orderNumber) implements Command<Boolean> {
+public record SetAwaitingValidationOrderStatusCommand(Long orderNumber) implements Command<Boolean> {
 }
