+++ services/ordering/src/main/java/com/eshop/ordering/api/application/commands/SetPaidOrderStatusCommand.java
@@ -2,5 +2,5 @@ package com.eshop.ordering.api.application.commands;
 
 import an.awesome.pipelinr.Command;
 
-public record SetPaidOrderStatusCommand(Integer orderNumber) implements Command<Boolean> {
+public record SetPaidOrderStatusCommand(Long orderNumber) implements Command<Boolean> {
 }
