+++ services/ordering/src/main/java/com/eshop/ordering/api/application/commands/SetStockConfirmedOrderStatusCommand.java
@@ -2,5 +2,5 @@ package com.eshop.ordering.api.application.commands;
 
 import an.awesome.pipelinr.Command;
 
-public record SetStockConfirmedOrderStatusCommand(Integer orderNumber) implements Command<Boolean> {
+public record SetStockConfirmedOrderStatusCommand(Long orderNumber) implements Command<Boolean> {
 }
