+++ services/ordering/src/main/java/com/eshop/ordering/api/application/commands/ShipOrderCommand.java
@@ -2,5 +2,5 @@ package com.eshop.ordering.api.application.commands;
 
 import an.awesome.pipelinr.Command;
 
-public record ShipOrderCommand(Integer orderNumber) implements Command<Boolean> {
+public record ShipOrderCommand(Long orderNumber) implements Command<Boolean> {
 }
